Only rescale images that are larger than the input dimension
Smaller images are not enlarged, just padded. All padding is zero-padding. This is following the suggestions in Hashemi 2019. Input dimension was changed to 128, as about 82% of training images have a long dimensions that is <= 128 px. 64 principal components preserve about 80% of the variance in the training data, so 64 PCs are kept.